+++ blosc/delta.h
@@ -12,10 +12,10 @@
 #include <stdio.h>
 #include <stdint.h>
 
-void delta_encoder(const uint8_t* dref, size_t offset, int32_t nbytes,
-                   size_t typesize, const uint8_t* src, uint8_t* dest);
+void delta_encoder(const uint8_t* dref, int32_t offset, int32_t nbytes,
+                   int32_t typesize, const uint8_t* src, uint8_t* dest);
 
-void delta_decoder(const uint8_t* dref, size_t offset, size_t nbytes,
-                   size_t typesize, uint8_t* dest);
+void delta_decoder(const uint8_t* dref, int32_t offset, int32_t nbytes,
+                   int32_t typesize, uint8_t* dest);
 
 #endif //BLOSC_DELTA_H
